requirements.txt
RealtimeSTT==0.3.100
ffmpeg-python==0.2.0
numpy>=1.24
sounddevice==0.4.6
pytest==8.2.0
prometheus-client>=0.16.0
